updated readmen
